// src/app/feature/profile/models/profile.models.ts
export interface IProfileForm {
  id: number;
  firstname: string;
  lastname: string;
  gender: string;
  designation: string;
  countryCode?: number;
  mobile: number;
  email: string;
  displayname: string;
  areaCode?: number;
  landline: number;
  dateOfBirth: string;
  country: string;
  state: string;
  city: string;
  street: string;
  pincode: string;
  entities: string;
  imgData: string;
}